Refactor: 불필요한 줄바꿈제거
