+++ src/Router.js
@@ -10,10 +10,10 @@ import PhoneCheck from './pages/signup/PhoneCheck';
 import SignUp from './pages/signup/SignUp';
 import SignUpPage from './pages/signup/SignUpPage';
 import DetailPage from './pages/detailpage/DetailPage';
-import Reservation2 from './pages/reservation/Reservation2';
+import Reservation from './pages/reservation/Reservation';
 import Points from './pages/mypage/components/Points';
 import Coupons from './pages/mypage/components/Coupons';
-import Reservation from './pages/mypage/components/Reservation';
+import ReservationList from './pages/mypage/components/ReservationList';
 import My from './pages/mypage/components/My';
 import Footer from './component/footer/Footer';
 
@@ -32,11 +32,11 @@ const Router = (props) => {
           <Route path='/signup-check' element={<PhoneCheck />} />
           <Route path='/signup-page' element={<SignUpPage />} />
           <Route path='/detailPage' element={<DetailPage />} />
-          <Route path='/reservation2' element={<Reservation2 />} />
+          <Route path='/reservation' element={<Reservation />} />
           <Route path='/accommodation/:type' element={<AccommodationList />} />
           <Route path='/mypage' element={<Points />} />
           <Route path='/coupons' element={<Coupons />} />
-          <Route path='/reservations' element={<Reservation />} />
+          <Route path='/reservation-list' element={<ReservationList />} />
           <Route path='/my' element={<My />} />
         </Routes>
         <Footer />
